+++ sol-did/client/packages/cli/src/commands/resolve/index.ts
@@ -48,7 +48,11 @@ export default class Resolve extends Command {
 
     const didSolIdentifier = DidSolIdentifier.parse(args.didsol);
 
-    const service = await DidSolService.build(didSolIdentifier);
+    const service = DidSolService.build(didSolIdentifier);
+
+    // console.log(`did:sol: ${service.did}`);
+    // console.log(`did:sol PDA: ${service.didDataAccount.toBase58()}`);
+    // console.log(`did:sol legacy PDA: ${service.legacyDidDataAccount.toBase58()}`);
 
     const doc = await service.resolve();
     this.log(JSON.stringify(doc, null, 2));