+++ sol-did/tests/suite-alloc/alloc.ts
@@ -40,9 +40,9 @@ describe('sol-did alloc operations', () => {
   let didDataAccountSize: number = 0;
 
   before(async () => {
-    [didData, didDataPDABump] = await findProgramAddress(authority.publicKey);
+    [didData, didDataPDABump] = findProgramAddress(authority.publicKey);
 
-    service = await DidSolService.buildFromAnchor(
+    service = DidSolService.buildFromAnchor(
       program,
       DidSolIdentifier.create(authority.publicKey, TEST_CLUSTER),
       programProvider